pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["DOBRO-228 <matematikaigor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
psycopg2 = "^2.9.1"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.1"
django-bootstrap4 = "^3.0.1"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.1"
wemake-python-styleguide = "^0.15.3"
isort = "^5.9.3"
pytest = "^6.2.5"
django-filter = "^21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
